price-reporter: move client to pkg, add per-network constructors with default url
