+++ packages/price-reporter/package.json
@@ -1,0 +1,39 @@
+{
+    "name": "@renegade-fi/price-reporter",
+    "version": "0.4.11",
+    "description": "A TypeScript client for interacting with the Price Reporter",
+    "private": true,
+    "files": [
+        "dist/**",
+        "!dist/**/*.tsbuildinfo",
+        "src/**/*.ts",
+        "!src/**/*.test.ts",
+        "!src/**/*.test-d.ts"
+    ],
+    "sideEffects": false,
+    "type": "module",
+    "main": "./dist/index.js",
+    "types": "./dist/types/index.d.ts",
+    "typings": "./dist/types/index.d.ts",
+    "exports": {
+        ".": {
+            "types": "./dist/types/index.d.ts",
+            "default": "./dist/index.js"
+        },
+        "./package.json": "./package.json"
+    },
+    "scripts": {
+        "build": "pnpm run clean && pnpm run build:esm+types",
+        "build:esm+types": "tsc --project tsconfig.build.json --outDir ./dist --declaration --declarationMap --declarationDir ./dist/types",
+        "clean": "rm -rf dist tsconfig.tsbuildinfo",
+        "typecheck": "tsc --noEmit"
+    },
+    "keywords": [],
+    "author": "",
+    "dependencies": {
+        "axios": "^1.6.8",
+        "@renegade-fi/core": "workspace:^",
+        "@renegade-fi/token": "workspace:^",
+        "neverthrow": "^8.2.0"
+    }
+}
