+++ packages/price-reporter/src/index.ts
@@ -1,0 +1,1 @@
+export { PriceReporterClient } from "./client.js";
