+++ packages/price-reporter/tsconfig.build.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "include": ["src/**/*.ts"],
+    "exclude": ["src/**/*.test.ts", "src/**/*.test-d.ts"],
+    "compilerOptions": {
+        "sourceMap": true
+    }
+}
