+++ packages/price-reporter/tsconfig.json
@@ -1,0 +1,5 @@
+{
+    "extends": "./tsconfig.build.json",
+    "include": ["src/**/*.ts", "test/**/*.ts"],
+    "exclude": []
+}
